requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.11
boto3>=1.34.0
anthropic>=0.21.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pandas>=2.2.0
numpy>=1.26.0
pytest>=8.0.0